Add files via upload
Under the constraints of a limited dataset, our research project investigates and emphasises the
outcomes of generative AI and the CNN model’s accuracy in an array of image classification tasks. An
enriched dataset incorporating neural style transfer, the original cat and dog images, and their augmented
variants using OpenAI’s DALL-E are utilised in a comparative analysis alongside three distinct CNN
models trained on separate datasets. The outcomes demonstrate the effectiveness of generative AI in
augmenting data. The model with the most augmentations attained the highest accuracy on the test
set, at 91.86 percent, representing a 4.16 percent advance over the model trained on the original dataset.
Innovative AI methods improve CNN performance and generalizability, especially in confined data setting,
according to the study.